Ravinder | Merge Settings back button functionality in beta

+++ apps/web-giddh/src/app/shared/header/header.component.ts
@@ -334,17 +334,17 @@ export class HeaderComponent implements OnInit, AfterViewInit, OnDestroy, AfterV
         // SETTING CURRENT PAGE ON ROUTE CHANGE
         this.router.events.pipe(takeUntil(this.destroyed$)).subscribe(event => {
             if (event instanceof NavigationStart) {
-                if(this.router.url.includes("/pages/settings") && !this.generalService.getSessionStorage("previousPage")) {
+                if((this.router.url.includes("/pages/settings") || this.router.url.includes("/pages/user-details")) && !this.generalService.getSessionStorage("previousPage")) {
                     this.generalService.setSessionStorage("previousPage", this.currentPageUrl);
                 }
 
-                if(!this.router.url.includes("/pages/settings") && this.generalService.getSessionStorage("previousPage")) {
+                if(!this.router.url.includes("/pages/settings") && !this.router.url.includes("/pages/user-details") && this.generalService.getSessionStorage("previousPage")) {
                     this.generalService.removeSessionStorage("previousPage");
                 }
                 this.addClassInBodyIfPageHasTabs();
             }
             if (event instanceof NavigationEnd) {
-                if(!this.router.url.includes("/pages/settings")) {
+                if(!this.router.url.includes("/pages/settings") && !this.router.url.includes("/pages/user-details")) {
                     this.currentPageUrl = this.router.url;
                 }
                 this.setCurrentPage();
